+++ src/main/java/org/thoughtcrime/securesms/AllMediaGalleryFragment.java
@@ -66,6 +66,11 @@ public class AllMediaGalleryFragment
     this.noMedia      = ViewUtil.findById(view, R.id.no_images);
     this.gridManager  = new StickyHeaderGridLayoutManager(getCols());
 
+    // allow content to be drawn behind the navigation bar
+    recyclerView.setClipToPadding(false);
+    // add padding to avoid content hidden behind system bars
+    ViewUtil.applyWindowInsets(recyclerView, true, false, true, true);
+
     this.recyclerView.setAdapter(new AllMediaGalleryAdapter(getContext(),
                                                          GlideApp.with(this),
                                                          new BucketedThreadMediaLoader.BucketedThreadMedia(getContext()),